fixes rerendering bug
