+++ src/components/Button.svelte
@@ -5,6 +5,7 @@
 </script>
 
 <button
+  type="button"
   class="p-1 m-1 w-24 {selected && 'bg-gray-800 text-white border-gray-800'}"
   on:click={onClick}>{text}</button
 >
