+++ src/main.ts
@@ -4,7 +4,7 @@ import { getRandomColor } from './utils'
 const app = new App({
   target: document.body,
   props: {
-    color: getRandomColor()
+    base: getRandomColor()
   }
 })
