--- src/utils/getColors.ts
@@ -1,3 +1,0 @@
-import type { SquareDimensionProps } from '../types'
-
-export function getColors(squares: SquareDimensionProps[], options) {}
